수업내용 보완

+++ jinyoung_client/CObj.cpp
@@ -15,8 +15,8 @@ CObj::~CObj()
 void CObj::render(HDC _dc)
 {
 	Rectangle(_dc
-		, m_Pos.x - m_Scale.x / 2
-		, m_Pos.y - m_Scale.y / 2
-		, m_Pos.x + m_Scale.x / 2
-		, m_Pos.y + m_Scale.y / 2);
+		, int(m_Pos.x - m_Scale.x / 2)
+		, int(m_Pos.y - m_Scale.y / 2)
+		, int(m_Pos.x + m_Scale.x / 2)
+		, int(m_Pos.y + m_Scale.y / 2));
 }
